pyproject.toml
[tool.poetry]
name = "yert"
version = "0.1.0"
description = ""
authors = ["AbstractUmbra <Umbra@AbstractUmbra.xyz>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = {git = "https://github.com/Rapptz/discord.py", rev = "master"}
discord-ext-menus = {git = "https://github.com/Rapptz/discord-ext-menus", rev = "master"}
jishaku = {git = "https://github.com/Gorialis/jishaku/", rev = "master"}
asyncpg = "*"
async_cleverbot = "*"
async_cse = "*"
bs4 = "*"
fuzzywuzzy = "*"
jikanpy = "*"
lxml = "*"
numpy = "*"
nhentaio = "*"
psutil = "*"
rule34 = "*"
Pillow = "*"
googletrans = "3.1.0a0"

[tool.poetry.dev-dependencies]
black = "*"
flake8 = "*"
"discord.py-stubs" = "^1.5.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
